Rename sqlc package from `db` to `dbsqlc`
Another small-ish one in which we rename the sqlc package from `db` to
`dbsqlc` to match the convention found in the main River project. IMO
this naming is a bit better because it's more suggestive when reading
code (and not already familiar with the project) that the calls are
related to sqlc, and gives a more distinctive search string when trying
to find symbols in code.
